+++ src/cmd/script/script_callbacks.cpp
@@ -61,7 +61,7 @@ string varToString (varInst * vi) {
     if (vi->objectname=="string") {
       return *((string *)vi->object);
     }else {
-      return XMLSupport::tostring ((int)vi->object);
+      return XMLSupport::tostring ((long)vi->object);
     }
   }
 }
